fix(ft_split.c): changed the index type to size_t

+++ ft_strncmp.c
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	index;
+	size_t	index;
 
 	index = 0;
 	while (index < n && s1[index] && s1[index] == s2[index])
